learning-coach-mcp/pyproject.toml
[project]
name = "learning-coach-mcp"
version = "0.1.0"
description = "AI Learning Coach MCP Server - Personalized learning assistant using Claude Memory and RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "supabase>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "ragas>=0.1.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
